refactor: move pbdata decode into module
